requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
joblib
numpy
pandas
python-dotenv
supabase
apscheduler